Remove IGamemode, create BaseGamemode instead
The core game loop (as far as I know) is going to be the same across all gamemodes. Because of this, it's better to just have a single base class cover 99% of the functionality rather than re-implementing across multiple instances of IGamemode.
