--- code/States/Gamemodes/IGamemode.cs
@@ -1,10 +1,0 @@
-﻿namespace Grubs.States;
-
-public interface IGamemode
-{
-	bool UsedTurn { get; }
-	TimeUntil TimeUntilTurnEnd { get; }
-
-	void UseTurn();
-	void NextTurn();
-}
